manga-translator/requirements.txt
# ─── Core numerics / vision ─────────────────────────────────────────────────────
numpy==1.26.4          # imgaug needs <2.0 (uses np.sctypes)  :contentReference[oaicite:1]{index=1}
opencv-python==4.11.0.86     # 2025-04-10 release :contentReference[oaicite:1]{index=1}
pillow==11.2.1               # 2025-04-12 release :contentReference[oaicite:2]{index=2}
# GPU stack
torch>=2.6,<2.8
torchvision>=0.20,<0.23
torchaudio>=2.5,<2.8
timm==1.0.15                 # transformer vision models :contentReference[oaicite:5]{index=5}
kornia==0.8.1                # vision ops (LaMa needs gaussian_blur2d) :contentReference[oaicite:6]{index=6}
einops==0.8.0                # tensor rearrange util   :contentReference[oaicite:7]{index=7}
albumentations==0.5.2        # last release that keeps DualIAATransform & IAA* ops
scikit-image==0.23.2         # image utils (optional but handy) :contentReference[oaicite:9]{index=9}

# ─── In-painting & detector helpers ─────────────────────────────────────────────
ultralytics==8.3.151         # YOLOv8 CLI & Python API :contentReference[oaicite:10]{index=10}
manga-ocr==0.1.14            # latest JP manga OCR model :contentReference[oaicite:11]{index=11}

# ─── LLM / translation stack ───────────────────────────────────────────────────
transformers==4.52.4         # HF models :contentReference[oaicite:12]{index=12}
sentencepiece==0.2.0         # tokenizer runtime :contentReference[oaicite:13]{index=13}
argostranslate==1.9.6        # offline NMT engine   :contentReference[oaicite:14]{index=14}
ctranslate2==4.6.0           # fast CPU/GPU inference (argos dependency)
googletrans==4.0.2           # async Google-translate wrapper :contentReference[oaicite:15]{index=15}
deepl==1.22.0                # DeepL official SDK (2025-04-30) :contentReference[oaicite:16]{index=16}
azure-ai-translation-text==1.0.1   # Azure Translator (optional) :contentReference[oaicite:17]{index=17}
azure-core==1.34.0                 # Azure SDK core lib :contentReference[oaicite:18]{index=18}
ollama==0.5.1                      # local-LLM client (fixes follow_redirects) :contentReference[oaicite:19]{index=19}

# ─── Web / UI front-ends ───────────────────────────────────────────────────────
streamlit==1.45.0            # GUI app framework   :contentReference[oaicite:20]{index=20}
gradio==5.33.0               # alt GUI & demos   :contentReference[oaicite:21]{index=21}

# ─── Networking & utilities ────────────────────────────────────────────────────
httpx==0.27.2                # pin for ollama/googletrans compat :contentReference[oaicite:22]{index=22}
requests==2.32.3             # HTTP classic   :contentReference[oaicite:23]{index=23}
easydict==1.13               # attr-style dict (LaMa) :contentReference[oaicite:24]{index=24}
tqdm==4.66.4                 # nice progress bars

# ─── optional quality-of-life extras ───────────────────────────────────────────
rich==13.7.1                 # pretty tracebacks / logs
hydra-core==1.3.2                 # config framework LaMa uses :contentReference[oaicite:0]{index=0}
scikit-learn==1.4.2          # required by saicinpainting evaluation
pytorch-lightning==2.2.4   # ✔ installs & provides import pytorch_lightning
webdataset==0.2.111            # needed by saicinpainting.datasets
